refactor(habitat): modularize D2 registry and Grit adapter
