+++ .grit/patterns/habitat/checks/habitat_grit_adapter_domain_paths.md
@@ -1,0 +1,53 @@
+---
+level: error
+---
+# Habitat Grit Adapter Domain Paths
+
+The Grit adapter is a generic tool boundary. Repo or product scan roots belong
+in registry metadata, not hard-coded adapter source.
+
+```grit
+language js(typescript)
+
+`$body` where {
+  $filename <: r".*tools/habitat-harness/src/adapters/grit/.*\.ts$",
+  $text = text($body),
+  or {
+    $text <: includes "packages",
+    $text <: includes "apps/",
+    $text <: includes "mods/",
+    $text <: includes "mods/mod-swooper-maps",
+    $text <: includes "apps/mapgen-studio",
+    $text <: includes ".civ7"
+  }
+}
+```
+
+## Matches fixture
+
+```typescript
+// @filename: tools/habitat-harness/src/adapters/grit/constants.ts
+export const roots = ["mods/mod-swooper-maps/src/domain"];
+
+// @filename: tools/habitat-harness/src/adapters/grit/constants.ts
+export const studio = "apps/mapgen-studio/src";
+
+// @filename: tools/habitat-harness/src/adapters/grit/constants.ts
+export const packages = ["packages"];
+
+// @filename: tools/habitat-harness/src/adapters/grit/constants.ts
+export const packageRoot = "packages/civ7-adapter";
+
+// @filename: tools/habitat-harness/src/adapters/grit/constants.ts
+export const resources = ".civ7/outputs/resources";
+```
+
+## Ignores fixture
+
+```typescript
+// @filename: tools/habitat-harness/src/adapters/grit/constants.ts
+export const injectedProbeRoot = "tools/habitat-harness/injected-probe-roots";
+
+// @filename: tools/habitat-harness/src/rules/rules.json
+const registryMetadataCanNameProductRoots = "mods/mod-swooper-maps/src/domain";
+```
