+++ docs/projects/habitat-harness/openspec-remediation/packet-index.md
@@ -16,7 +16,7 @@ it does not repeat local worktree paths or branch names.
 | --- | --- | --- | --- | --- | --- |
 | D0 | D0 Command Surface Inventory | `deep-habitat-d0-command-surface-inventory` | Fresh worktree from main, dependency install, matrix validation, OpenSpec validation, D0 command/test probes | All later dominoes | implementation matrix authored at `docs/projects/habitat-harness/public-surface-compatibility-matrix.md` with 328 concrete rows; packet-boundary approved for D1 implementation-start preparation |
 | D1 | D1 Receipt Contract Boundary | `deep-habitat-d1-receipt-contract-boundary` | D0 command surface inventory; concrete D0 matrix rows exist; D1 execution inventory, output-family citations, inherited D0 evidence baseline, and implementation-start review dispositions cleared before source implementation | D6, D7, D8, D9, D10, D11, D12, D13, D14 | source implementation submitted in PR #1835 and PR #1836; command-engine monolith deleted; focused command modules and verify receipt failed-state boundary implemented; no unresolved accepted D1 P1/P2 blockers recorded |
-| D2 | D2 Rule Registry Metadata Contract | `deep-habitat-d2-rule-registry-metadata-contract` | D0, D1; concrete D0 matrix rows and D1 malformed-metadata output-family citations required before source implementation | D3, D4, D5, D6, D7, D8, D10, D13 | accepted for design/specification; final review and code/topology recheck found no unresolved P1/P2 blockers; not implementation-complete |
+| D2 | D2 Rule Registry Metadata Contract | `deep-habitat-d2-rule-registry-metadata-contract` | D0, D1; concrete D0 matrix rows and D1 malformed-metadata output-family citations required before source implementation | D3, D4, D5, D6, D7, D8, D10, D13 | source implementation active; TypeBox registry parser and named projections are live; focused D2 parser/projection/consumer gates pass; structural adapter-domain enforcement is registered as GritQL; final packet approval remains open |
 | D3 | D3 Workspace Graph Boundary | `deep-habitat-d3-workspace-graph-boundary` | D0, D2; concrete D0 public-surface rows and live D2 `ruleGraphFacts` implementation required before source implementation | D4, D7, D12 | accepted for design/specification; final rereview found no unresolved P1/P2 blockers; not implementation-complete |
 | D4 | D4 Classify Orientation And Routing | `deep-habitat-d4-orientation-routing` | D0, D1 vocabulary, D2, D3; concrete D0 rows plus live D2/D3 implementation facts required before source implementation | D14 | accepted for design/specification; final domain/ontology, OpenSpec/testing, and topology/cross-domino rereviews found no unresolved P1/P2 blockers; not implementation-complete |
 | D5 | D5 Baseline Authority | `deep-habitat-d5-baseline-authority` | D0, D2; concrete D0 rows and live D2 `ruleBaselineFacts`/baseline projections required before source implementation | D7, D8 | accepted for design/specification; final domain/ontology, OpenSpec/testing, and topology/TypeScript/cross-domino rereviews found no unresolved P1/P2 blockers; not implementation-complete |
