+++ openspec/changes/deep-habitat-d2-rule-registry-metadata-contract/workstream/downstream-realignment-ledger.md
@@ -2,20 +2,20 @@
 
 ## Status
 
-D2 downstream handoffs are accepted for design/specification after final D2 rereview and focused code/topology recheck. Downstream packet design may consume accepted D2 projections; source implementation that depends on live projections waits for D2 implementation, concrete D0 matrix rows, and D1 malformed-metadata output-family citations.
+D2 downstream handoffs are accepted for design/specification after final D2 rereview and focused code/topology recheck. D2 source implementation now provides live TypeBox-backed registry parsing and named projection functions for downstream packets. Downstream packet source implementation still waits on each packet's own gates, D0/D1 citations, and any non-D2 owner packets named below.
 
 ## Direct Consumers
 
 | Downstream | D2 projection consumed | Design status after D2 acceptance | Source implementation gate |
 | --- | --- | --- | --- |
-| D3 Workspace Graph Boundary | `ruleGraphFacts` | May design graph ownership, target availability, and Nx graph behavior against structured D2 graph declarations. | Blocked until D2 implementation provides live graph projections and D0 rows cover Nx target metadata. |
-| D4 Orientation And Routing | `ruleRoutingFacts` plus D3 graph results | May design classify/orientation scenarios against D2 path coverage states. | Blocked until D2 removes prose-scope routing authority and D3 graph truth exists where needed. |
-| D5 Baseline Authority | `ruleBaselineFacts` | May define baseline load/shrink/growth/debt states against D2 baseline relations. | Blocked until D2 implementation removes permissive whole-row baseline parsing. |
-| D6 Diagnostic Pattern Catalog | `ruleGritFacts`, diagnostic identifiers, execution adapter ids | May define diagnostic catalog and adapter failure taxonomy against D2 Grit facts. | Blocked until D2 implementation provides Grit projections and removes pattern-id fallback. |
-| D7 Structural Enforcement Pipeline | `ruleSelectorFacts`, `ruleReportFacts`, `ruleCommandExecutionFacts`, `ruleRoutingFacts`, `ruleBaselineFacts`, `ruleGritFacts`, `ruleFileLayerFacts` | May design enforcement aggregation against named consumer projections and D1 output families. | Blocked until D2 implementation provides live projections and D5/D6/D10 owner packets close their own source behavior. |
-| D8 Pattern Governance | `ruleGovernanceFacts`, `ruleGritFacts`, `ruleBaselineFacts` | May design Pattern Authority lifecycle/admission against D2 registry references. | Blocked until D2 implementation projects governance facts and D5 baseline relation behavior is available. |
-| D10 Protected Zone Authority | `ruleFileLayerFacts` plus G-HOST declarations and D1 refusals | May design guard/refusal behavior against D2 generated-zone references. | Blocked until D2 implementation provides generated-zone projections and G-HOST accepts host-policy boundary. |
-| D13 Scaffolding And Refusal Contracts | `ruleGovernanceFacts`, `ruleFileLayerFacts`, scaffold relation metadata | May design generator/refusal scenarios against D2 registry references and D8/G-HOST decisions. | Blocked until D2/D8/G-HOST implementation facts exist. |
+| D3 Workspace Graph Boundary | `ruleGraphFacts` | May design graph ownership, target availability, and Nx graph behavior against structured D2 graph declarations. | D2 graph projection is live; D3 still owns resolved graph truth and D0 rows for Nx target metadata. |
+| D4 Orientation And Routing | `ruleRoutingFacts` plus D3 graph results | May design classify/orientation scenarios against D2 path coverage states. | D2 routing projection is live and prose-scope routing authority is removed; D4 still waits for D3 graph truth where needed. |
+| D5 Baseline Authority | `ruleBaselineFacts` | May define baseline load/shrink/growth/debt states against D2 baseline relations. | D2 baseline projection is live for current registry state; D5 still owns shrink/growth/debt policy. |
+| D6 Diagnostic Pattern Catalog | `ruleGritFacts`, diagnostic identifiers, execution adapter ids | May define diagnostic catalog and adapter failure taxonomy against D2 Grit facts. | D2 Grit projection is live and pattern-id fallback is removed; D6 still owns diagnostic catalog semantics. |
+| D7 Structural Enforcement Pipeline | `ruleSelectorFacts`, `ruleReportFacts`, `ruleCommandExecutionFacts`, `ruleRoutingFacts`, `ruleBaselineFacts`, `ruleGritFacts`, `ruleFileLayerFacts` | May design enforcement aggregation against named consumer projections and D1 output families. | D2 projections are live; D7 still waits for D5/D6/D10 owner packets where their behavior is consumed. |
+| D8 Pattern Governance | `ruleGovernanceFacts`, `ruleGritFacts`, `ruleBaselineFacts` | May design Pattern Authority lifecycle/admission against D2 registry references. | D2 governance projection is live and contains registry relation facts only; D8 still owns lifecycle/admission/fixture authority. |
+| D10 Protected Zone Authority | `ruleFileLayerFacts` plus G-HOST declarations and D1 refusals | May design guard/refusal behavior against D2 generated-zone references. | D2 file-layer projection is live; G-HOST/D10 still own host and protected-zone policy. |
+| D13 Scaffolding And Refusal Contracts | `ruleGovernanceFacts`, `ruleFileLayerFacts`, scaffold relation metadata | May design generator/refusal scenarios against D2 registry references and D8/G-HOST decisions. | D2 governance/file-layer facts are live; D8/G-HOST/D13 still own admission and scaffold/refusal behavior. |
 
 ## Indirect Consumers
 
@@ -42,14 +42,13 @@ D2 downstream handoffs are accepted for design/specification after final D2 rere
 
 ## Required Index Alignment
 
-- Keep D2 status as `accepted for design/specification; final review and code/topology recheck found no unresolved P1/P2 blockers; implementation active; not implementation-complete`.
+- Keep D2 status as `source implementation active; focused D2 projection/parser gates pass; structural adapter-domain enforcement is covered by a registered GritQL rule; final review/Graphite closure still open`.
 - D2 source implementation requires concrete D0 matrix rows and D1 malformed-metadata citations; those implementation-start citations now live in `workstream/implementation-start-inventory.md`.
 - G-HOST is not listed as D2-enabled. G-HOST is parallel host-policy work requiring D0 and D1; D10 consumes both D2 and G-HOST.
 - D10 remains the packet where D2 generated-zone facts and G-HOST host-policy declarations meet.
 
 ## Non-Claims
 
-- This ledger accepts only D2 design/specification handoffs.
-- This ledger does not implement projections.
+- This ledger records D2 design/specification handoffs and the current live projection implementation status.
 - This ledger does not accept, repair, or close any downstream packet.
-- This ledger does not authorize source implementation from D2 design alone; implementation proceeds through the implementation-start inventory and TypeBox-first registry boundary.
+- This ledger does not authorize downstream source implementation from D2 alone; downstream packets still proceed through their own implementation-start inventories and owner gates.
