+++ tools/habitat-harness/baselines/grit-habitat-adapter-domain-paths.json
@@ -1,0 +1,1 @@
+[]
