+++ tools/habitat-harness/src/adapters/grit/constants.ts
@@ -1,0 +1,25 @@
+export const gritBin = "grit";
+
+export const injectedProbeRoot = "tools/habitat-harness/injected-probe-roots";
+
+export const protectedScanRootPrefixes = [
+  ".git/",
+  ".grit/cache/",
+  "dist/",
+  "node_modules/",
+  "tools/habitat-harness/dist/",
+];
+
+export const docsLocalCheckoutPathsRewritePattern =
+  ".grit/patterns/habitat/apply/docs_local_checkout_paths_rewrite.md";
+
+export const gritCandidateExtensions = new Set([
+  ".cjs",
+  ".cts",
+  ".js",
+  ".jsx",
+  ".mjs",
+  ".mts",
+  ".ts",
+  ".tsx",
+]);
