+++ tools/habitat-harness/src/adapters/grit/index.ts
@@ -1,0 +1,35 @@
+export {
+  injectedProbeRoot,
+} from "./constants.js";
+export {
+  discoverGritScanRoots,
+  effectiveGritScanRoots,
+  validateScanRoots,
+} from "./scan-roots/index.js";
+export {
+  parseGritCheckOutput,
+  parseGritCheckTextOutput,
+} from "./output/index.js";
+export {
+  projectGritResults,
+} from "./projection.js";
+export {
+  gritCheckProgram,
+  gritCheckRequest,
+} from "./request.js";
+export {
+  resetGritCacheForTests,
+  runGritRule,
+  runGritRules,
+} from "./runner.js";
+export type {
+  GritCheckCacheMode,
+  GritCheckOptions,
+  GritCheckOutputFormat,
+  GritCheckParseResult,
+  GritCheckRequestOptions,
+  GritProjectionOptions,
+  GritReport,
+  GritResult,
+} from "./types.js";
+export type { GritScanRootValidationOptions } from "./scan-roots/index.js";
