+++ tools/habitat-harness/src/rules/registry/schema.js
@@ -1,0 +1,1 @@
+export * from "./schema.ts";
